+++ helpers/migration.ts
@@ -1,4 +1,7 @@
 import type {ObjectSchemaDefinition} from '../schema';
+import type {SetEndpoint} from '../types';
+import type {SetEndpointDef} from '../types';
+import {ensureExists} from '../helpers/ensure';
 
 export function objectSchemaHelper<T extends ObjectSchemaDefinition<string, string>>(schema: T): ObjectSchemaHelper<T> {
   return new ObjectSchemaHelper(schema);
@@ -33,5 +36,37 @@ export class ObjectSchemaHelper<T extends ObjectSchemaDefinition<string, string>
 
   get type() {
     return this._schema.type;
+  }
+}
+
+export function setEndpointHelper(step: SetEndpoint) {
+  return new SetEndpointHelper(step);
+}
+
+export class SetEndpointHelper {
+  private readonly _step: SetEndpoint;
+
+  constructor(step: SetEndpoint) {
+    this._step = step;
+  }
+
+  get getOptions() {
+    return ensureExists(this._step.getOptions ?? this._step.getOptionsFormula);
+  }
+}
+
+export function setEndpointDefHelper(step: SetEndpointDef) {
+  return new SetEndpointDefHelper(step);
+}
+
+export class SetEndpointDefHelper {
+  private readonly _step: SetEndpointDef;
+
+  constructor(step: SetEndpointDef) {
+    this._step = step;
+  }
+
+  get getOptions() {
+    return ensureExists(this._step.getOptions ?? this._step.getOptionsFormula);
   }
 }