+++ types.ts
@@ -159,16 +159,20 @@ export interface SetEndpoint {
    * `{display: '<display name>', value: '<endpoint>'}` if wanting to render a display
    * label to the user rather than rendering the underlying value directly.
    */
-  getOptionsFormula: MetadataFormula;
+  getOptions?: MetadataFormula; // TODO(packs-dev): Make this required after migration.
+  /** @deprecated Use {@link getOptions} */
+  getOptionsFormula?: MetadataFormula;
 }
 
 /**
  * Simplified configuration for {@link SetEndpoint} that a pack developer can specify when calling
  * {@link setUserAuthentication} or {@link setSystemAuthentication}.
  */
-export type SetEndpointDef = Omit<SetEndpoint, 'getOptionsFormula'> & {
+export type SetEndpointDef = Omit<SetEndpoint, 'getOptions' | 'getOptionsFormula'> & {
+  /** See {@link SetEndpoint.getOptions} */
+  getOptions?: MetadataFormulaDef;
   /** See {@link SetEndpoint.getOptionsFormula} */
-  getOptionsFormula: MetadataFormulaDef;
+  getOptionsFormula?: MetadataFormulaDef;
 };
 
 /**